+++ demo/demos/position.tsx
@@ -31,11 +31,17 @@ const Component = defineComponent({
       >
         <select v-model={this.position} placeholder="Select Position">
           <option value="left">Left</option>
+          <option value="left-start">Left Start</option>
+          <option value="left-end">Left End</option>
           <option value="right">Right</option>
+          <option value="right-start">Right Start</option>
+          <option value="right-end">Right End</option>
           <option value="top">Top</option>
+          <option value="top-start">Top Start</option>
+          <option value="top-end">Top End</option>
           <option value="bottom">Bottom</option>
-          <option value="top-right">Top Right</option>
-          <option value="bottom-right">Bottom Right</option>
+          <option value="bottom-start">Bottom Start</option>
+          <option value="bottom-end">Bottom End</option>
         </select>
         <br />
         <br />